# build: Share core bpftrace code with test binary