--- a/src/CMakeLists.txt
+++ b/src/CMakeLists.txt
@@ -11,7 +11,7 @@ else()
   set(BPFTRACE bpftrace)
 endif()
 
-add_executable(${BPFTRACE}
+add_library(libbpftrace
   attached_probe.cpp
   bpffeature.cpp
   bpforc.cpp
@@ -21,7 +21,6 @@ add_executable(${BPFTRACE}
   clang_parser.cpp
   disasm.cpp
   driver.cpp
-  fake_map.cpp
   lockdown.cpp
   log.cpp
   map.cpp
@@ -31,15 +30,21 @@ add_executable(${BPFTRACE}
   procmon.cpp
   printf.cpp
   resolve_cgroupid.cpp
-  signal.cpp
   struct.cpp
   tracepoint_format_parser.cpp
   types.cpp
   usdt.cpp
   utils.cpp
   ${BFD_DISASM_SRC}
+)
+# So it's not "liblibbpftrace"
+set_target_properties(libbpftrace PROPERTIES PREFIX "")
+
+add_executable(${BPFTRACE}
   ${MAIN_SRC}
 )
+
+target_link_libraries(${BPFTRACE} libbpftrace)
 
 if (BUILD_FUZZ)
   target_compile_options(${BPFTRACE} PUBLIC "-DFUZZ")
@@ -63,94 +68,94 @@ if (BUILD_FUZZ)
 endif()
 
 if(HAVE_NAME_TO_HANDLE_AT)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_NAME_TO_HANDLE_AT=1)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_NAME_TO_HANDLE_AT=1)
 endif(HAVE_NAME_TO_HANDLE_AT)
 if(HAVE_BCC_PROG_LOAD)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_BCC_PROG_LOAD)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_BCC_PROG_LOAD)
 endif(HAVE_BCC_PROG_LOAD)
 if(HAVE_BCC_CREATE_MAP)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_BCC_CREATE_MAP)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_BCC_CREATE_MAP)
 endif(HAVE_BCC_CREATE_MAP)
 if(HAVE_BCC_ELF_FOREACH_SYM)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_BCC_ELF_FOREACH_SYM)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_BCC_ELF_FOREACH_SYM)
 endif(HAVE_BCC_ELF_FOREACH_SYM)
 if(HAVE_BCC_USDT_ADDSEM)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_BCC_USDT_ADDSEM)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_BCC_USDT_ADDSEM)
 endif(HAVE_BCC_USDT_ADDSEM)
 if (LIBBPF_BTF_DUMP_FOUND)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_LIBBPF_BTF_DUMP)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_LIBBPF_BTF_DUMP)
   target_include_directories(${BPFTRACE} PUBLIC ${LIBBPF_INCLUDE_DIRS})
-  target_link_libraries(${BPFTRACE} ${LIBBPF_LIBRARIES})
+  target_link_libraries(libbpftrace ${LIBBPF_LIBRARIES})
   if (HAVE_LIBBPF_BTF_DUMP_EMIT_TYPE_DECL)
-    target_compile_definitions(${BPFTRACE} PRIVATE HAVE_LIBBPF_BTF_DUMP_EMIT_TYPE_DECL)
+    target_compile_definitions(libbpftrace PRIVATE HAVE_LIBBPF_BTF_DUMP_EMIT_TYPE_DECL)
   endif()
 endif(LIBBPF_BTF_DUMP_FOUND)
 if (LIBBPF_FOUND)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_LIBBPF)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_LIBBPF)
 endif(LIBBPF_FOUND)
 
 if (HAVE_LIBBPF_MAP_BATCH)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_LIBBPF_MAP_BATCH)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_LIBBPF_MAP_BATCH)
 endif()
 
 if (HAVE_BCC_KFUNC)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_BCC_KFUNC)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_BCC_KFUNC)
 endif(HAVE_BCC_KFUNC)
 if(HAVE_BFD_DISASM)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_BFD_DISASM)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_BFD_DISASM)
   if(LIBBFD_DISASM_FOUR_ARGS_SIGNATURE)
-    target_compile_definitions(${BPFTRACE} PRIVATE LIBBFD_DISASM_FOUR_ARGS_SIGNATURE)
+    target_compile_definitions(libbpftrace PRIVATE LIBBFD_DISASM_FOUR_ARGS_SIGNATURE)
   endif(LIBBFD_DISASM_FOUR_ARGS_SIGNATURE)
   if(STATIC_LINKING)
     add_library(LIBBFD STATIC IMPORTED)
     set_property(TARGET LIBBFD PROPERTY IMPORTED_LOCATION ${LIBBFD_LIBRARIES})
-    target_link_libraries(${BPFTRACE} LIBBFD)
+    target_link_libraries(libbpftrace LIBBFD)
     add_library(LIBOPCODES STATIC IMPORTED)
     set_property(TARGET LIBOPCODES PROPERTY IMPORTED_LOCATION ${LIBOPCODES_LIBRARIES})
-    target_link_libraries(${BPFTRACE} LIBOPCODES)
+    target_link_libraries(libbpftrace LIBOPCODES)
     add_library(LIBIBERTY STATIC IMPORTED)
     set_property(TARGET LIBIBERTY PROPERTY IMPORTED_LOCATION ${LIBIBERTY_LIBRARIES})
-    target_link_libraries(${BPFTRACE} LIBIBERTY)
+    target_link_libraries(libbpftrace LIBIBERTY)
   else()
-    target_link_libraries(${BPFTRACE} ${LIBBFD_LIBRARIES})
-    target_link_libraries(${BPFTRACE} ${LIBOPCODES_LIBRARIES})
+    target_link_libraries(libbpftrace ${LIBBFD_LIBRARIES})
+    target_link_libraries(libbpftrace ${LIBOPCODES_LIBRARIES})
   endif(STATIC_LINKING)
 endif(HAVE_BFD_DISASM)
 if(LIBBCC_ATTACH_KPROBE_SIX_ARGS_SIGNATURE)
-  target_compile_definitions(${BPFTRACE} PRIVATE LIBBCC_ATTACH_KPROBE_SIX_ARGS_SIGNATURE)
+  target_compile_definitions(libbpftrace PRIVATE LIBBCC_ATTACH_KPROBE_SIX_ARGS_SIGNATURE)
 endif(LIBBCC_ATTACH_KPROBE_SIX_ARGS_SIGNATURE)
 if(LIBBCC_ATTACH_UPROBE_SEVEN_ARGS_SIGNATURE)
-  target_compile_definitions(${BPFTRACE} PRIVATE LIBBCC_ATTACH_UPROBE_SEVEN_ARGS_SIGNATURE)
+  target_compile_definitions(libbpftrace PRIVATE LIBBCC_ATTACH_UPROBE_SEVEN_ARGS_SIGNATURE)
 endif(LIBBCC_ATTACH_UPROBE_SEVEN_ARGS_SIGNATURE)
 
 if (ALLOW_UNSAFE_PROBE)
-  target_compile_definitions(${BPFTRACE} PRIVATE HAVE_UNSAFE_PROBE)
+  target_compile_definitions(libbpftrace PRIVATE HAVE_UNSAFE_PROBE)
 endif(ALLOW_UNSAFE_PROBE)
 
-target_link_libraries(${BPFTRACE} arch ast parser resources)
+target_link_libraries(libbpftrace arch ast parser resources)
 
-target_link_libraries(${BPFTRACE} ${LIBBCC_LIBRARIES})
+target_link_libraries(libbpftrace ${LIBBCC_LIBRARIES})
 if(STATIC_LINKING)
   # These are not part of the static libbcc so have to be added separate
-  target_link_libraries(${BPFTRACE} ${LIBBCC_BPF_LIBRARY_STATIC})
-  target_link_libraries(${BPFTRACE} ${LIBBPF_LIBRARIES})
-  target_link_libraries(${BPFTRACE} ${LIBBCC_LOADER_LIBRARY_STATIC})
+  target_link_libraries(libbpftrace ${LIBBCC_BPF_LIBRARY_STATIC})
+  target_link_libraries(libbpftrace ${LIBBPF_LIBRARIES})
+  target_link_libraries(libbpftrace ${LIBBCC_LOADER_LIBRARY_STATIC})
 
   add_library(LIBELF STATIC IMPORTED)
   set_property(TARGET LIBELF PROPERTY IMPORTED_LOCATION ${LIBELF_LIBRARIES})
-  target_link_libraries(${BPFTRACE} LIBELF)
+  target_link_libraries(libbpftrace LIBELF)
 
   if(NOT STATIC_LIBC)
     set_target_properties(${BPFTRACE} PROPERTIES LINK_FLAGS "${EMBEDDED_LINK_FLAGS}")
   endif()
 else()
-  target_link_libraries(${BPFTRACE} ${LIBELF_LIBRARIES})
+  target_link_libraries(libbpftrace ${LIBELF_LIBRARIES})
 endif(STATIC_LINKING)
 
 # Support for std::filesystem
 # GCC version <9 and Clang (all versions) require -lstdc++fs
 if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang" OR ${CMAKE_CXX_COMPILER_VERSION} VERSION_LESS "9")
-  target_link_libraries(${BPFTRACE} "stdc++fs")
+  target_link_libraries(libbpftrace "stdc++fs")
 endif()
 
 if (BUILD_ASAN)
@@ -180,7 +185,7 @@ install(TARGETS ${BPFTRACE} DESTINATION ${CMAKE_INSTALL_BINDIR})
 set(KERNEL_HEADERS_DIR "" CACHE PATH "Hard-code kernel headers directory")
 if (KERNEL_HEADERS_DIR)
   MESSAGE(STATUS "Using KERNEL_HEADERS_DIR=${KERNEL_HEADERS_DIR}")
-  target_compile_definitions(${BPFTRACE} PUBLIC KERNEL_HEADERS_DIR="${KERNEL_HEADERS_DIR}")
+  target_compile_definitions(libbpftrace PUBLIC KERNEL_HEADERS_DIR="${KERNEL_HEADERS_DIR}")
 endif()
 
 execute_process(