--- a/src/ast/CMakeLists.txt
+++ b/src/ast/CMakeLists.txt
@@ -1,12 +1,14 @@
 add_library(ast
   ast.cpp
   attachpoint_parser.cpp
+  callback_visitor.cpp
   codegen_llvm.cpp
+  fake_map.cpp
   field_analyser.cpp
   irbuilderbpf.cpp
   printer.cpp
   semantic_analyser.cpp
-  callback_visitor.cpp
+  signal.cpp
   visitors.cpp
 )