--- a/tests/CMakeLists.txt
+++ b/tests/CMakeLists.txt
@@ -45,88 +45,10 @@ add_executable(bpftrace_test
   utils.cpp
 
   ${CODEGEN_SRC}
-
-  ${CMAKE_SOURCE_DIR}/src/attached_probe.cpp
-  ${CMAKE_SOURCE_DIR}/src/bpftrace.cpp
-  ${CMAKE_SOURCE_DIR}/src/bpffeature.cpp
-  ${CMAKE_SOURCE_DIR}/src/bpforc.cpp
-  ${CMAKE_SOURCE_DIR}/src/btf.cpp
-  ${CMAKE_SOURCE_DIR}/src/child.cpp
-  ${CMAKE_SOURCE_DIR}/src/clang_parser.cpp
-  ${CMAKE_SOURCE_DIR}/src/disasm.cpp
-  ${CMAKE_SOURCE_DIR}/src/driver.cpp
-  ${CMAKE_SOURCE_DIR}/src/fake_map.cpp
-  ${CMAKE_SOURCE_DIR}/src/log.cpp
-  ${CMAKE_SOURCE_DIR}/src/map.cpp
-  ${CMAKE_SOURCE_DIR}/src/mapkey.cpp
-  ${CMAKE_SOURCE_DIR}/src/output.cpp
-  ${CMAKE_SOURCE_DIR}/src/printf.cpp
-  ${CMAKE_SOURCE_DIR}/src/probe_matcher.cpp
-  ${CMAKE_SOURCE_DIR}/src/procmon.cpp
-  ${CMAKE_SOURCE_DIR}/src/resolve_cgroupid.cpp
-  ${CMAKE_SOURCE_DIR}/src/signal.cpp
-  ${CMAKE_SOURCE_DIR}/src/struct.cpp
-  ${CMAKE_SOURCE_DIR}/src/tracepoint_format_parser.cpp
-  ${CMAKE_SOURCE_DIR}/src/types.cpp
-  ${CMAKE_SOURCE_DIR}/src/usdt.cpp
-  ${CMAKE_SOURCE_DIR}/src/utils.cpp
-  ${BFD_DISASM_SRC}
 )
 
-if (HAVE_LIBBPF_MAP_BATCH)
-  target_compile_definitions(bpftrace_test PRIVATE HAVE_LIBBPF_MAP_BATCH)
-endif()
-
-if(HAVE_NAME_TO_HANDLE_AT)
-  target_compile_definitions(bpftrace_test PRIVATE HAVE_NAME_TO_HANDLE_AT=1)
-endif(HAVE_NAME_TO_HANDLE_AT)
-if(HAVE_BCC_PROG_LOAD)
-  target_compile_definitions(bpftrace_test PRIVATE HAVE_BCC_PROG_LOAD)
-endif(HAVE_BCC_PROG_LOAD)
-if(HAVE_BCC_CREATE_MAP)
-  target_compile_definitions(bpftrace_test PRIVATE HAVE_BCC_CREATE_MAP)
-endif(HAVE_BCC_CREATE_MAP)
-if (LIBBPF_BTF_DUMP_FOUND)
-  target_compile_definitions(bpftrace_test PRIVATE HAVE_LIBBPF_BTF_DUMP)
-  target_include_directories(bpftrace_test PUBLIC ${LIBBPF_INCLUDE_DIRS})
-  target_link_libraries(bpftrace_test ${LIBBPF_LIBRARIES})
-  if (HAVE_LIBBPF_BTF_DUMP_EMIT_TYPE_DECL)
-    target_compile_definitions(bpftrace_test PRIVATE HAVE_LIBBPF_BTF_DUMP_EMIT_TYPE_DECL)
-  endif()
-endif(LIBBPF_BTF_DUMP_FOUND)
-if (HAVE_BCC_KFUNC)
-  target_compile_definitions(bpftrace_test PRIVATE HAVE_BCC_KFUNC)
-endif(HAVE_BCC_KFUNC)
-if(HAVE_BCC_USDT_ADDSEM)
-  target_compile_definitions(bpftrace PRIVATE HAVE_BCC_USDT_ADDSEM)
-endif(HAVE_BCC_USDT_ADDSEM)
 target_compile_definitions(bpftrace_test PRIVATE TEST_CODEGEN_LOCATION="${CMAKE_SOURCE_DIR}/tests/codegen/llvm/")
-if(HAVE_BFD_DISASM)
-  target_compile_definitions(bpftrace_test PRIVATE HAVE_BFD_DISASM)
-  if(LIBBFD_DISASM_FOUR_ARGS_SIGNATURE)
-    target_compile_definitions(bpftrace_test PRIVATE LIBBFD_DISASM_FOUR_ARGS_SIGNATURE)
-  endif(LIBBFD_DISASM_FOUR_ARGS_SIGNATURE)
-  if(STATIC_LINKING)
-    add_library(LIBBFD STATIC IMPORTED)
-    set_property(TARGET LIBBFD PROPERTY IMPORTED_LOCATION ${LIBBFD_LIBRARIES})
-    target_link_libraries(bpftrace_test LIBBFD)
-    add_library(LIBOPCODES STATIC IMPORTED)
-    set_property(TARGET LIBOPCODES PROPERTY IMPORTED_LOCATION ${LIBOPCODES_LIBRARIES})
-    target_link_libraries(bpftrace_test LIBOPCODES)
-    add_library(LIBIBERTY STATIC IMPORTED)
-    set_property(TARGET LIBIBERTY PROPERTY IMPORTED_LOCATION ${LIBIBERTY_LIBRARIES})
-    target_link_libraries(bpftrace_test LIBIBERTY)
-  else()
-    target_link_libraries(bpftrace_test ${LIBBFD_LIBRARIES})
-    target_link_libraries(bpftrace_test ${LIBOPCODES_LIBRARIES})
-  endif(STATIC_LINKING)
-endif(HAVE_BFD_DISASM)
-if(LIBBCC_ATTACH_KPROBE_SIX_ARGS_SIGNATURE)
-  target_compile_definitions(bpftrace_test PRIVATE LIBBCC_ATTACH_KPROBE_SIX_ARGS_SIGNATURE)
-endif(LIBBCC_ATTACH_KPROBE_SIX_ARGS_SIGNATURE)
-if(LIBBCC_ATTACH_UPROBE_SEVEN_ARGS_SIGNATURE)
-  target_compile_definitions(bpftrace_test PRIVATE LIBBCC_ATTACH_UPROBE_SEVEN_ARGS_SIGNATURE)
-endif(LIBBCC_ATTACH_UPROBE_SEVEN_ARGS_SIGNATURE)
+target_link_libraries(bpftrace_test libbpftrace)
 
 if(CMAKE_SYSTEM_PROCESSOR STREQUAL "aarch64")
   target_compile_definitions(bpftrace_test PRIVATE ARCH_AARCH64)
@@ -140,34 +62,14 @@ elseif(CMAKE_SYSTEM_PROCESSOR STREQUAL "x86_64")
   target_compile_definitions(bpftrace_test PRIVATE ARCH_X86_64)
 endif()
 
-target_link_libraries(bpftrace_test arch ast parser resources)
-
-target_link_libraries(bpftrace_test ${LIBBCC_LIBRARIES})
-if (STATIC_LINKING)
+if(STATIC_LINKING)
   if(EMBED_LLVM OR EMBED_CLANG)
     set_target_properties(bpftrace_test PROPERTIES LINK_FLAGS "${EMBEDDED_LINK_FLAGS}")
   endif()
   if(STATIC_LIBC)
     set_target_properties(bpftrace_test PROPERTIES LINK_FLAGS "-static")
   endif()
-
-  # These are not part of the static libbcc so have to be added separate
-  target_link_libraries(bpftrace_test ${LIBBCC_BPF_LIBRARY_STATIC})
-  target_link_libraries(bpftrace_test ${LIBBPF_LIBRARIES})
-  target_link_libraries(bpftrace_test ${LIBBCC_LOADER_LIBRARY_STATIC})
-
-  add_library(LIBELF STATIC IMPORTED)
-  set_property(TARGET LIBELF PROPERTY IMPORTED_LOCATION ${LIBELF_LIBRARIES})
-  target_link_libraries(bpftrace_test LIBELF)
-else()
-  target_link_libraries(bpftrace_test ${LIBELF_LIBRARIES})
 endif(STATIC_LINKING)
-
-# Support for std::filesystem
-# GCC version <9 and Clang (all versions) require -lstdc++fs
-if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang" OR ${CMAKE_CXX_COMPILER_VERSION} VERSION_LESS "9")
-  target_link_libraries(bpftrace_test "stdc++fs")
-endif()
 
 find_package(Threads REQUIRED)